Move funcs to util.
